+++ R/FuncUserUncor.R
@@ -1,6 +1,6 @@
 ## Function to call cpbayes_uncor (CPBayes function for uncorrelated phenotypes)
 #' Run uncorrelated version of CPBayes.
-#' 
+#'
 #' Run uncorrelated version of CPBayes when the main genetic effect (beta/log(odds ratio)) estimates across
 #'  studies/traits are uncorrelated.
 #' @param BetaHat A numeric vector of length K where K is the number of phenotypes. It
@@ -14,8 +14,8 @@
 #'  Default is `Variant'.
 #' @param UpdateSlabVar A logical vector of length one. If TRUE, the variance of the slab distribution
 #'  that presents the prior distribution of non-null effects
-#'  is updated at each MCMC iteration in a range (MinSlabVar -- MaxSlabVar) (see next). If FALSE, 
-#'   it is fixed at (MinSlabVar + MaxSlabVar)/2. Default is TRUE. 
+#'  is updated at each MCMC iteration in a range (MinSlabVar -- MaxSlabVar) (see next). If FALSE,
+#'   it is fixed at (MinSlabVar + MaxSlabVar)/2. Default is TRUE.
 #' @param MinSlabVar A numeric value greater than 0.01 providing the minimum value of
 #'  the variance of the slab distribution. Default is 0.6.
 #' @param MaxSlabVar A numeric value smaller than 10.0 providing the maximum value of
@@ -27,15 +27,15 @@
 #'    can experiment different choices of these three arguments: UpdateSlabVar, MinSlabVar, and MaxSlabVar.
 #' @param MCMCiter A positive integer greater than or equal to 7,000 providing the total number of
 #'  iterations in the MCMC. Default is 20,000.
-#' @param Burnin A positive integer greater than or equal to 2,000 providing the burn in period 
+#' @param Burnin A positive integer greater than or equal to 2,000 providing the burn in period
 #' in the MCMC. Default is 5,000. Note that the MCMC sample size (MCMCiter - Burnin) must be at least 5,000.
-#' @return The output produced by the function is a list which consists of various components. 
+#' @return The output produced by the function is a list which consists of various components.
 #'    \item{variantName}{It is the name of the genetic variant provided by the user. If not
-#'     specified by the user, default name is `Variant'.} 
+#'     specified by the user, default name is `Variant'.}
 #'    \item{log10_BF}{It provides the log10(Bayes factor) produced by CPBayes that measures the
 #'     evidence of the overall pleiotropic association.}
 #'    \item{locFDR}{It provides the local false discovery rate (posterior probability of null association) produced by CPBayes
-#'     (a Bayesian analog of the p-value) which is a measure of the evidence of the 
+#'      which is a measure of the evidence of the
 #'     aggregate-level pleiotropic association. Bayes factor is adjusted for prior odds, but
 #'      locFDR is solely a function of the posterior odds. locFDR can sometimes be small
 #'      indicating an association, but log10_BF may not indicate an association. Hence, always check both log10_BF and locFDR.}
@@ -44,11 +44,11 @@
 #'    \item{important_traits}{It provides the traits which yield a trait-specific posterior probability of
 #'     association (PPAj) > 20\%. Even if a phenotype is not selected in the optimal subset of non-null
 #'      traits, it can produce a non-negligible value of trait-specific posterior probability of
-#'       association (PPAj). Note that, `important_traits' is expected to include the traits 
+#'       association (PPAj). Note that, `important_traits' is expected to include the traits
 #'       already contained in `subset'. It provides both the name of the important traits and
 #'        their corresponding values of PPAj. Always check 'important_traits' even if 'subset' contains
 #'         a single trait. It helps to better explain an observed pleiotropic signal.}
-#'    \item{auxi_data}{It contains supplementary data including the MCMC data which is used later 
+#'    \item{auxi_data}{It contains supplementary data including the MCMC data which is used later
 #'     by \code{\link{post_summaries}} and \code{\link{forest_cpbayes}}:
 #'        \enumerate{
 #'            \item traitNames: Name of all the phenotypes.
@@ -63,11 +63,11 @@
 #'    }
 #'    \item{runtime}{It provides the runtime (in seconds) taken by \code{\link{cpbayes_uncor}}. It will help the user
 #'     to plan the whole analysis.}
-#' 
+#'
 #' @references Majumdar A, Haldar T, Bhattacharya S, Witte JS (2018) An efficient Bayesian meta analysis approach for studying cross-phenotype genetic associations. PLoS Genet 14(2): e1007139.
-#' 
-#' @seealso \code{\link{post_summaries}}, \code{\link{forest_cpbayes}}, \code{\link{cpbayes_cor}}, \code{\link{estimate_corln}}
-#' 
+#'
+#' @seealso \code{\link{analytic_locFDR_BF_uncor}}, \code{\link{post_summaries}}, \code{\link{forest_cpbayes}}, \code{\link{analytic_locFDR_BF_cor}}, \code{\link{cpbayes_cor}}, \code{\link{estimate_corln}}
+#'
 #' @examples
 #' data(ExampleDataUncor)
 #' BetaHat <- ExampleDataUncor$BetaHat
@@ -78,11 +78,11 @@
 #' SNP1 <- "rs1234"
 #' result <- cpbayes_uncor(BetaHat, SE, Phenotypes = traitNames, Variant = SNP1)
 #' str(result)
-#' 
+#'
 #' @export
 cpbayes_uncor <- function(BetaHat, SE, Phenotypes, Variant, UpdateSlabVar = TRUE, MinSlabVar = 0.6, MaxSlabVar = 1.0, MCMCiter = 20000, Burnin = 5000)
 {
-  
+
   UpdateDE <- UpdateSlabVar
   # Check whether any of the primary arguments is missing
     if(missing(BetaHat) || missing (SE))
@@ -115,7 +115,7 @@ cpbayes_uncor <- function(BetaHat, SE, Phenotypes, Variant, UpdateSlabVar = TRUE
     # Check whether argument 5 is a vector of length 1
     if(!is.vector(UpdateDE) || (length(UpdateDE) != 1))
     {
-      warning("UpdateSlabVar is not a vector of length 1 (default option used).", call. = FALSE)
+      warning("UpdateSlabVar is not a scalar (default option used).", call. = FALSE)
       UpdateDE <- TRUE
     }
     if(!is.logical(UpdateDE))
@@ -125,111 +125,17 @@ cpbayes_uncor <- function(BetaHat, SE, Phenotypes, Variant, UpdateSlabVar = TRUE
     }
 
   # Argument 6 & 7:: Minimum and maximum value of slab variance
-    MinSlabVarDefault <- 0.6
-    MaxSlabVarDefault <- 1.0
-    MinSlabVarBound <- 0.01
-    MaxSlabVarBound <- 10.0
-    
-    # Check whether argument 6 is a vector of length 1
-    if(!is.vector(MinSlabVar) || (length(MinSlabVar) != 1))
-    {
-      warning("MinSlabVar is not a vector of length 1 (default option used).", call. = FALSE)
-      MinSlabVar <- MinSlabVarDefault
-    }
-    # Check whether argument 6 is numeric
-    if(!is.numeric(MinSlabVar)){
-      warning("MinSlabVar is not numeric (default option used).", call. = FALSE)
-      MinSlabVar <- MinSlabVarDefault
-    }
-    # Check whether argument 6 is more than its minimum bound
-    if(MinSlabVar < MinSlabVarBound){
-      warning("MinSlabVar should not be smaller than ", MinSlabVarBound, ", so it is assigned to ", MinSlabVarBound, ".", call. = FALSE)
-      MinSlabVar <- MinSlabVarBound
-    }
+    SlabVarList <- chkSlabVar(MinSlabVar, MaxSlabVar)
+    MinSlabVar <- SlabVarList[["MinSlabVar"]]
+    MaxSlabVar <- SlabVarList[["MaxSlabVar"]]
 
-  # Argument 7 :: Maximum value of slab variance    
-    
-    # Check whether argument 7 is a vector of length 1
-    if(!is.vector(MaxSlabVar) || (length(MaxSlabVar) != 1))
-    {
-      warning("MaxSlabVar is not a vector of length 1 (default option used).", call. = FALSE)
-      MaxSlabVar <- MaxSlabVarDefault
-    }
-    # Check whether argument 7 is numeric
-    if(!is.numeric(MaxSlabVar)){
-      warning("MaxSlabVar is not numeric (default option used).", call. = FALSE)
-      MaxSlabVar <- MaxSlabVarDefault
-    }
-    # Check whether argument 7 is less than a maximum bound
-    if(MaxSlabVar > MaxSlabVarBound){
-      warning("MaxSlabVar should not be bigger than ", MaxSlabVarBound, ", so it is assigned to ", MaxSlabVarBound, ".", call. = FALSE)
-      MaxSlabVar <- MaxSlabVarBound
-    }
-    
-  # Argument 6 and 7 :: Checking MinSlabVar < MaxSlabVar
-    if(MinSlabVar >= MaxSlabVar){
-      warning("MaxSlabVar is not bigger than MinSlabVar! (default option used).", call. = FALSE)
-      MinSlabVar <- MinSlabVarDefault
-      MaxSlabVar <- MaxSlabVarDefault
-    }
-    
 	# Argument 8 :: Number of MCMC iteration
-    mcmcDefault <- 20000
-    # Check whether argument 8 is a vector of length 1
-    if(!is.vector(MCMCiter) || (length(MCMCiter) != 1))
-    {
-      warning("MCMCiter is not a vector of length 1 (default option used).", call. = FALSE)
-      MCMCiter <- mcmcDefault
-    }
-		# Check whether argument 8 is numeric and integer
-		if(!is.numeric(MCMCiter) || MCMCiter%%1 != 0)
-		{
-		  warning("MCMCiter not provided as integer (default option used).", call. = FALSE)
-		  MCMCiter <- mcmcDefault
-		}
-		# Check whether argument 8 is more than 7000
-		if(MCMCiter < 7000)
-		{
-		  warning("MCMCiter should be at least 7000 (default option used).", call. = FALSE)
-		  MCMCiter <- mcmcDefault
-		}
-			
-	# Argument 9 :: Burnin
-    BurninDefault <- 5000
-    # Check whether argument 9 is a vector of length 1
-    if(!is.vector(Burnin) || (length(Burnin) != 1))
-    {
-      warning("Burnin is not a vector of length 1 (default option used).", call. = FALSE)
-      Burnin <- BurninDefault
-    }
-		# Check whether argument 9 is numeric and integer
-		if(!is.numeric(Burnin) || Burnin%%1 != 0)
-		{
-		  warning("Burnin not provided as integer (default option used).", call. = FALSE)
-		  Burnin <- BurninDefault
-		}
-		# Check whether argument 9 is more than 2000
-		if(Burnin < 2000)
-		{
-		  warning("Burnin should be at least 2000 (default option used).", call. = FALSE)
-		  Burnin <- BurninDefault
-		}
-    
-  # Argument 8 and 9
-    if((MCMCiter - Burnin) < 5000)
-    {
-      warning("MCMC sample size (MCMCiter - Burnin) provided less than 5000 (default options used)", call. = FALSE)
-      MCMCiter <- mcmcDefault
-      Burnin <- BurninDefault
-    }
-    
+    mcmcParam <- chkMCMCparam(MCMCiter, Burnin)
+    MCMCiter <- mcmcParam[["MCMCiter"]]
+    Burnin <- mcmcParam[["Burnin"]]
 	# Call CPBayes function
 	RESULT <- CPBayes_uncor( variantName, Phenotypes, BetaHat, SE, UpdateDE, MinSlabVar, MaxSlabVar, MCMCiter, Burnin )
 	print_result(RESULT)
 	invisible(RESULT)
 }
-
-
-
-