+++ tests/testthat/test-InputRestCor.R
@@ -6,8 +6,8 @@ test_that("Throws warning if MCMCiter parameter is not a integer greater than a 
   expect_warning(cpbayes_cor(1:10, 1:10, ExampleDataCor$cor, MCMCiter = "123"), "MCMCiter not provided as integer*")
   expect_warning(cpbayes_cor(1:10, 1:10, ExampleDataCor$cor, MCMCiter = 123.5), "MCMCiter not provided as integer*")
   expect_warning(cpbayes_cor(1:10, 1:10, ExampleDataCor$cor, MCMCiter = -1), "MCMCiter should be at least 7000*")
-  expect_warning(cpbayes_cor(1:10, 1:10, ExampleDataCor$cor, MCMCiter = c(10000, 20000)), "MCMCiter is not a vector of length 1*")
-  expect_warning(cpbayes_cor(1:10, 1:10, ExampleDataCor$cor, MCMCiter = as.matrix(c(10000, 20000))), "MCMCiter is not a vector of length 1*")
+  expect_warning(cpbayes_cor(1:10, 1:10, ExampleDataCor$cor, MCMCiter = c(10000, 20000)), "MCMCiter is not a scalar*")
+  expect_warning(cpbayes_cor(1:10, 1:10, ExampleDataCor$cor, MCMCiter = as.matrix(c(10000, 20000))), "MCMCiter is not a scalar*")
 })
 
 test_that("Throws warning if Burnin parameter is not a integer lower than a cutoff", {
@@ -16,8 +16,8 @@ test_that("Throws warning if Burnin parameter is not a integer lower than a cuto
   expect_warning(cpbayes_cor(1:10, 1:10, ExampleDataCor$cor, Burnin = "123"), "Burnin not provided as integer*")
   expect_warning(cpbayes_cor(1:10, 1:10, ExampleDataCor$cor, Burnin = 123.5), "Burnin not provided as integer*")
   expect_warning(cpbayes_cor(1:10, 1:10, ExampleDataCor$cor, Burnin = -1), "Burnin should be at least 2000*")
-  expect_warning(cpbayes_cor(1:10, 1:10, ExampleDataCor$cor, Burnin = c(10000, 20000)), "Burnin is not a vector of length 1*")
-  expect_warning(cpbayes_cor(1:10, 1:10, ExampleDataCor$cor, Burnin = as.matrix(c(10000, 20000))), "Burnin is not a vector of length 1*")
+  expect_warning(cpbayes_cor(1:10, 1:10, ExampleDataCor$cor, Burnin = c(10000, 20000)), "Burnin is not a scalar*")
+  expect_warning(cpbayes_cor(1:10, 1:10, ExampleDataCor$cor, Burnin = as.matrix(c(10000, 20000))), "Burnin is not a scalar*")
 })
 
 test_that("Throws warning if MCMC sample size is less than 5000", {